+++ src/header.rs
@@ -7,7 +7,7 @@ use embedded_graphics::prelude::*;
 use nom::{
     bytes::complete::tag,
     combinator::map_opt,
-    number::complete::{le_u16, le_u32},
+    number::complete::{le_i32, le_u16, le_u32},
     IResult,
 };
 
@@ -25,6 +25,22 @@ pub enum Bpp {
     Bits24,
     /// 32 bits per pixel.
     Bits32,
+}
+
+/// Image row order
+#[derive(Copy, Clone, Eq, PartialEq, Ord, PartialOrd, Hash, Debug)]
+#[non_exhaustive]
+pub enum RowOrder {
+    /// Bottom-up (standard)
+    BottomUp,
+    /// Top-down
+    TopDown,
+}
+
+impl Default for RowOrder {
+    fn default() -> Self {
+        Self::BottomUp
+    }
 }
 
 impl Bpp {
@@ -75,6 +91,9 @@ pub struct Header {
 
     /// Bit masks for the color channels.
     pub channel_masks: Option<ChannelMasks>,
+
+    /// Row order of the image data within the file
+    pub row_order: RowOrder,
 }
 
 impl Header {
@@ -86,11 +105,17 @@ impl Header {
         let (input, image_data_start) = le_u32(input)?;
         let (input, header_size) = le_u32(input)?;
         let (input, image_width) = le_u32(input)?;
-        let (input, image_height) = le_u32(input)?;
+        let (input, image_height) = le_i32(input)?;
         let (input, _color_planes) = le_u16(input)?;
         let (input, bpp) = Bpp::parse(input)?;
         let (input, compression_method) = CompressionMethod::parse(input)?;
         let (input, image_data_len) = le_u32(input)?;
+
+        let row_order = if image_height < 0 {
+            RowOrder::TopDown
+        } else {
+            RowOrder::BottomUp
+        };
 
         let (input, channel_masks) = if compression_method == CompressionMethod::Bitfields {
             // BMP header versions can be distinguished by the header length.
@@ -127,10 +152,11 @@ impl Header {
             Header {
                 file_size,
                 image_data_start: image_data_start as usize,
-                image_size: Size::new(image_width, image_height),
+                image_size: Size::new(image_width, image_height.abs() as u32),
                 image_data_len,
                 bpp,
                 channel_masks,
+                row_order,
             },
         ))
     }