+++ tests/chessboard-8px-24bit.rs
@@ -1,5 +1,5 @@
 use embedded_graphics::prelude::*;
-use tinybmp::{Bpp, Header, RawBmp};
+use tinybmp::{Bpp, Header, RawBmp, RowOrder};
 
 const DATA: &[u8] = include_bytes!("./chessboard-8px-24bit.bmp");
 
@@ -16,6 +16,7 @@ fn chessboard_8px_24bit() {
             image_size: Size::new(8, 8),
             image_data_len: 192,
             channel_masks: None,
+            row_order: RowOrder::BottomUp
         }
     );
 
@@ -38,6 +39,7 @@ fn chessboard_8px_24bit_truncated_iter() {
             image_size: Size::new(8, 8),
             image_data_len: 192,
             channel_masks: None,
+            row_order: RowOrder::BottomUp
         }
     );