+++ tests/chessboard-8px-color-16bit.rs
@@ -1,5 +1,5 @@
 use embedded_graphics::prelude::*;
-use tinybmp::{Bpp, ChannelMasks, Header, RawBmp};
+use tinybmp::{Bpp, ChannelMasks, Header, RawBmp, RowOrder};
 
 #[test]
 fn chessboard_8px_color_16bit() {
@@ -14,7 +14,8 @@ fn chessboard_8px_color_16bit() {
             bpp: Bpp::Bits16,
             image_size: Size::new(8, 8),
             image_data_len: 128,
-            channel_masks: Some(ChannelMasks::RGB565)
+            channel_masks: Some(ChannelMasks::RGB565),
+            row_order: RowOrder::BottomUp
         }
     );